+++ build.rs
@@ -22,5 +22,6 @@ fn main() {
         println!("cargo:rustc-cfg=feature=\"messagebuf\"");
         println!("cargo:rustc-cfg=feature=\"ovrhdr\"");
         println!("cargo:rustc-cfg=feature=\"subprio\"");
+        println!("cargo:rustc-cfg=feature=\"pi_mutex\"");
     }
 }